ordl_platform/backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ordl-platform-backend"
version = "0.1.0"
description = "ORDL clean-room fleet control plane backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "boto3>=1.35.0,<2.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
